Move Selection classes MushROMs project
* Move all classes that derive from ISelection to MushROMs project
folder.
* Add Tests.MushROMs project.
* Move SelectionTests class to Tests.MushROMs project.
