+++ src/Controls/TileMapControl.cs
@@ -8,6 +8,7 @@ using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Windows.Forms;
 using Helper;
+using MushROMs;
 
 namespace Controls
 {
