+++ src/Controls/TileMapControl1D.cs
@@ -6,6 +6,7 @@ using System;
 using System.Drawing;
 using System.Drawing.Drawing2D;
 using Helper;
+using MushROMs;
 
 namespace Controls
 {
