+++ src/MushROMs/ITileMap1D.cs
@@ -2,7 +2,7 @@
 //     Copyright (c) 2018 Nelson Garcia.
 // </copyright>
 
-namespace Helper
+namespace MushROMs
 {
     public interface ITileMap1D : ITileMap
     {
