+++ src/MushROMs/SingleSelection1D.cs
@@ -6,7 +6,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 
-namespace Helper
+namespace MushROMs
 {
     public sealed class SingleSelection1D : Selection1D
     {
