+++ src/MushROMs/SingleSelection2D.cs
@@ -1,20 +1,21 @@
-﻿// <copyright file="EmptySelection2D.cs" company="Public Domain">
+﻿// <copyright file="SingleSelection2D.cs" company="Public Domain">
 //     Copyright (c) 2018 Nelson Garcia.
 // </copyright>
 
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using Helper;
 
-namespace Helper
+namespace MushROMs
 {
-    internal sealed class EmptySelection2D : Selection2D
+    public sealed class SingleSelection2D : Selection2D
     {
         public override int Count
         {
             get
             {
-                return 0;
+                return 1;
             }
         }
 
@@ -22,34 +23,47 @@ namespace Helper
         {
             get
             {
-                throw new ArgumentOutOfRangeException(nameof(index));
+                if (index != 0)
+                {
+                    throw new ArgumentOutOfRangeException(nameof(index));
+                }
+
+                return StartPosition;
             }
         }
 
-        internal EmptySelection2D()
+        public SingleSelection2D(Position2D position)
         {
+            StartPosition = position;
         }
 
         public override Selection2D Copy()
         {
-            return Empty;
+            return new SingleSelection2D(StartPosition);
         }
 
         public override bool Contains(Position2D position)
         {
-            return false;
+            return position == StartPosition;
         }
 
         public override IEnumerator<Position2D> GetEnumerator()
         {
-            return new Enumerator();
+            return new Enumerator(StartPosition);
         }
 
         private struct Enumerator : IEnumerator<Position2D>
         {
+            private bool CanMove
+            {
+                get;
+                set;
+            }
+
             public Position2D Current
             {
                 get;
+                private set;
             }
 
             object IEnumerator.Current
@@ -60,12 +74,25 @@ namespace Helper
                 }
             }
 
+            public Enumerator(Position2D startPosition)
+            {
+                CanMove = true;
+                Current = startPosition;
+            }
+
             public void Reset()
             {
+                CanMove = true;
             }
 
             public bool MoveNext()
             {
+                if (CanMove)
+                {
+                    CanMove = false;
+                    return true;
+                }
+
                 return false;
             }
