+++ src/Tests.MushROMs/SelectionTests.cs
@@ -3,8 +3,8 @@
 // </copyright>
 
 using System.Collections.Generic;
-using Helper;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using MushROMs;
 
 namespace Tests.Helper
 {
